Refactor DnD, remove DungeonsAndDragons
Refactored the existing DnD class to include my content from the
DungeonsAndDragons class/locale. Also added .tool-versions to the
.gitignore so that asdf files won't be included on accident.
